requirements.txt
alembic==0.9.7
Babel==2.5.3
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pkg-resources==0.0.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2018.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
